Change: [NewGRF] Implement GRFv9.

+++ src/industrytype.h
@@ -130,7 +130,7 @@ struct IndustrySpec {
 	uint16_t callback_mask;                       ///< Bitmask of industry callbacks that have to be called
 	bool enabled;                               ///< entity still available (by default true).newgrf can disable it, though
 	GRFFileProps grf_prop;                      ///< properties related to the grf file
-	std::vector<uint8_t> random_sounds; ///< Random sounds;
+	std::vector<uint16_t> random_sounds; ///< Random sounds;
 
 	std::array<std::variant<CargoLabel, MixedCargoType>, INDUSTRY_ORIGINAL_NUM_OUTPUTS> produced_cargo_label; ///< Cargo labels of produced cargo for default industries.
 	std::array<std::variant<CargoLabel, MixedCargoType>, INDUSTRY_ORIGINAL_NUM_INPUTS> accepts_cargo_label; ///< Cargo labels of accepted cargo for default industries.
